Refactor own-side partially filled orders handling
If target spread is not reached and we need to place closer order, we need to
make sure current closest order is 100% unfilled. When target spread is
reached, we are replacing order only if it was filled no less than
`self.fill_threshold`. This helps to avoid too often replacements.