Change market center price handling in profit methods
Methods to store or get profit estimation data are depends on
get_market_center_price(). When there is no bids or asks,
get_market_center_price() will disable the whole worker. This is not
needed when we just want to store/calc profits, so we can use
`suppress_errors=True` here safely.

Closes: #563